feat(queries): divide arquivo de consultas graphql
estava ficando extenso e dificil de encontrar as coisas
