+++ src/views/apresentacao/Cliente.vue
@@ -118,7 +118,7 @@
 <script lang="ts">
 import Vue from "vue";
 import * as firebase from "firebase/app";
-import { GET_CLIENTE, UPDATE_CLIENTE } from "@/queries";
+import { GET_CLIENTE, UPDATE_CLIENTE } from "@/queries/cliente";
 import uuid from "uuid-random";
