+++ src/views/apresentacao/Usuario.vue
@@ -116,7 +116,7 @@
 <script lang="ts">
 import Vue from "vue";
 import * as firebase from "firebase/app";
-import { GET_USER, UPDATE_USER } from "@/queries";
+import { GET_USER, UPDATE_USER } from "@/queries/user";
 import uuid from "uuid-random";
 
 interface Usuario {
